pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "cybersecuritytools"
version = "0.1.0"
description = ""
authors = ["GDS Cyber Security"]
packages = [
    { include = "cybersecuritytools" },
    { include = "cybersecuritytools/**/*.py" },
]

[tool.poetry.dependencies]
boto3 = "*"
click = "*"
flask = "*"
oic = "*"
pyjwt = "*"
python = ">=3.6,<4.0"
requests = "*"
requests_toolbelt = "*"
serverless-wsgi = "*"
toml = "*"
dataclasses = { version = "*", python = "~3.6" }
boto3-stubs = {extras = ["essential", "logs"], version="*"}
splunk-sdk = "*"

[tool.poetry.dev-dependencies]
pytest = "*"
isort = "*"
black = { version = "*", allow-prereleases = true }
pytest-black = "*"
pytest-env = "*"
pytest-flake8 = "*"
pytest-isort = "*"
pytest-mock = "*"
pytest-mypy = "*"
pytest-cov = "*"
tox-gh-actions = "*"
moto = "*"
requests-mock = "*"

[tool.poetry.scripts]
cst = "cybersecuritytools.cst:cli"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--black --isort --cov=cybersecuritytools --cov-report=term --cov-report=html --doctest-modules --flake8 --mypy -vv"
testpaths = [
    "cybersecuritytools",
]
env =  [
    "AUTH_MODE=flask",
    # TODO: This could be hardcoded into the tests
    "SSM_PREFIX=/static-site/test",
    "STATIC_ROOT=./tests/static/",
    "LOG_LEVEL=ERROR"
]
